astramentor-backend/pyproject.toml
[tool.poetry]
name = "astramentor-backend"
version = "0.1.0"
description = "AI-powered Socratic tutoring backend for AstraMentor"
authors = ["AstraMentor Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
boto3 = "^1.34.34"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = "^5.0.1"
structlog = "^24.1.0"
tenacity = "^8.2.3"
circuitbreaker = "^2.0.0"
python-multipart = "^0.0.6"
langgraph = "^0.0.26"
langchain-core = "^0.1.23"
cryptography = "^42.0.0"
networkx = {version = ">=3.0,<4.0", python = ">=3.11,<3.14"}
tree-sitter = "^0.25.2"
tree-sitter-python = "^0.25.0"
tree-sitter-javascript = "^0.25.0"
tree-sitter-typescript = "^0.23.2"
tree-sitter-java = "^0.23.5"
tree-sitter-go = "^0.25.0"
tree-sitter-rust = "^0.24.0"
faiss-cpu = {version = ">=1.13.2,<2.0.0", python = ">=3.11,<3.15"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
hypothesis = "^6.98.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
